I decided to go ahead an upload what file I have from the project
I no longer have SolidWorks on any of my devices so I am not able to check if these are indeed MY parts but I hope they are.

Let me clarify, I may have been the inspector for the parts that are not mine here. This is because the parts needed to be verified before being used in the assembly.